Initial revision of card class
Created an abstract class to inherit to let me keep the existing classes as they are so it will build while I'm working.  That will probably be discarded once this original class can be removed.

This broke basically every card test, but once changed to the initializer, they start passing again.